sliderコンポーネントに一旦分割
Sliderは再実装したいねーーー
